// x-pack/platform/plugins/shared/alerting/server/task_runner/types.d.ts
export interface RuleTaskRunResult {
    state: RuleTaskState;
    schedule: IntervalSchedule | undefined;
    taskRunError?: DecoratedError;
    shouldDeleteTask?: boolean;
    shouldDisableTask?: boolean;
}
export declare const getDeleteRuleTaskRunResult: () => RuleTaskRunResult;
export interface RunRuleResult {
    metrics: RuleRunMetrics;
    state: RuleTaskState;
}
export interface RunRuleParams<Params extends RuleTypeParams> {
    apiKey: RawRule['apiKey'];
    uiamApiKey?: RawRule['uiamApiKey'];
    fakeRequest: KibanaRequest;
    rule: SanitizedRule<Params>;
    validatedParams: Params;
    version: string | undefined;
}
export interface RuleTaskInstance extends ConcreteTaskInstance {
    state: RuleTaskState;
}
// ...
export interface RuleRunnerErrorStackTraceLog {
    message: ElasticsearchError;
    stackTrace?: string;
}
export declare enum ApiKeyType {
    ES = "es",
    UIAM = "uiam"
}
export interface TaskRunnerContext {
    actionsConfigMap: ActionsConfigMap;
    actionsPlugin: ActionsPluginStartContract;
    alertsService: AlertsService | null;
    backfillClient: BackfillClient;
    basePathService: IBasePath;
    cancelAlertsOnRuleTimeout: boolean;
    connectorAdapterRegistry: ConnectorAdapterRegistry;
    data: DataPluginStart;
    dataViews: DataViewsPluginStart;
    elasticsearch: ElasticsearchServiceStart;
    encryptedSavedObjectsClient: EncryptedSavedObjectsClient;
    eventLogger: IEventLogger;
    executionContext: ExecutionContextStart;
    kibanaBaseUrl: string | undefined;
    logger: Logger;
    maintenanceWindowsService: MaintenanceWindowsService;
    maxAlerts: number;
    ruleTypeRegistry: RuleTypeRegistry;
    rulesSettingsService: RulesSettingsService;
    apiKeyType: ApiKeyType;
    savedObjects: SavedObjectsServiceStart;
    share: SharePluginStart;
    spaceIdToNamespace: SpaceIdToNamespaceFunction;
    uiSettings: UiSettingsServiceStart;
    usageCounter?: UsageCounter;
    getEventLogClient: (request: KibanaRequest) => IEventLogClient;
    isServerless: boolean;
    shouldGrantUiam?: boolean;
}